Expression Improvement

+++ function.cpp
@@ -434,10 +434,10 @@ bool FunctionC::runTruncdate(vector<string>* fieldnames, vector<string>* fieldva
     struct tm tm;
     long iSeconds = atol(sSeconds.c_str());
     if (strptime(sTm.c_str(), sFmt.c_str(), &tm)){
-      time_t t1 = ((long)mktime(&tm)/iSeconds)*iSeconds;
+      time_t t1 = ((long)((long)mktime(&tm)/iSeconds))*iSeconds;
       tm = *(localtime(&t1));
       sResult = dateToStr(tm, sFmt);
-      trace(DEBUG, "Truncating date '%s'(%d) to '%s'(%d), format:%s\n", sTm.c_str(), (long)mktime(&tm), sResult.c_str(), (long)t1, sFmt.c_str());
+      trace(DEBUG, "Truncating date '%s'(%u) to '%s'(%u), format:%s\n", sTm.c_str(), (long)mktime(&tm), sResult.c_str(), (long)t1, sFmt.c_str());
       return !sResult.empty();
     }else{
       trace(ERROR, "Failed to run truncdate(%s, %s)!\n", m_params[0].m_expStr.c_str(), m_params[1].m_expStr.c_str());
